[DEV-16075]: Rounded out management command setup and uses; updated Assistant model, migrations, and fixtures. TODO: Write tests.
