+++ usaspending_api/llm/fixtures/assistants.yaml
@@ -9,3 +9,4 @@
       topP: 1.0
       maxTokens: 5000
       stopSequences: []
+    is_active: true
